try to deal with simlinking nature of component library
because it's symlink the vue that is used inside of library get also
bundled with other application. So in result the bundle contain 2
versions of vue: one that is used inside of a component library consumer
and another one from component library itself.
`resolve.dedupe` should solve the issue so that listed
dependencies will always be taken from project root